+++ crates/flashblocks/src/lib.rs
@@ -8,10 +8,6 @@ pub(crate) use metrics::Metrics;
 
 mod pending_blocks;
 
-pub mod pubsub;
-
-pub mod rpc;
-
 pub mod state;
 
 pub mod subscription;
