+++ crates/rpc/Cargo.toml
@@ -11,6 +11,9 @@ repository.workspace = true
 workspace = true
 
 [dependencies]
+# Project
+base-reth-flashblocks.workspace = true
+
 # base/tips
 tips-core.workspace = true
 
@@ -22,19 +25,33 @@ reth-evm.workspace = true
 reth-optimism-evm.workspace = true
 reth-optimism-chainspec.workspace = true
 reth-transaction-pool = { workspace = true, features = ["test-utils"] }
+reth-rpc-eth-api.workspace = true
 
 # alloy
 alloy-primitives.workspace = true
 alloy-consensus.workspace = true
 alloy-eips.workspace = true
+alloy-rpc-types.workspace = true
+alloy-rpc-types-eth.workspace = true
+
+# op-alloy
+op-alloy-rpc-types.workspace = true
+op-alloy-network.workspace = true
+
+# tokio
+tokio.workspace = true
+tokio-stream.workspace = true
 
 # rpc
 jsonrpsee.workspace = true
+jsonrpsee-types.workspace = true
 
 # misc
 tracing.workspace = true
 eyre.workspace = true
 serde.workspace = true
+metrics.workspace = true
+metrics-derive.workspace = true
 
 [dev-dependencies]
 reth-optimism-primitives.workspace = true
@@ -51,8 +68,14 @@ reth-tracing.workspace = true
 reth-transaction-pool = { workspace = true, features = ["test-utils"] }
 serde_json.workspace = true
 tokio.workspace = true
+tokio-tungstenite.workspace = true
 base-reth-test-utils.workspace = true
 httpmock.workspace = true
+rollup-boost.workspace = true
+futures-util.workspace = true
+alloy-rpc-types-eth.workspace = true
+alloy-rpc-types-engine.workspace = true
+alloy-provider.workspace = true
 
 [package.metadata.cargo-udeps.ignore]
 normal = ["reth-optimism-cli"]
