+++ crates/rpc/src/base/mod.rs
@@ -1,5 +1,6 @@
 pub(crate) mod meter;
 pub(crate) mod meter_rpc;
+pub(crate) mod pubsub;
 pub(crate) mod traits;
 pub(crate) mod transaction_rpc;
 pub(crate) mod types;
